refactor: split app break-glass helpers
